Fix directory cleanup for e2e test on windows
