+++ dotnet/tools/Bzl/Files.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Diagnostics;
 using System.IO;
 
 namespace Bzl
@@ -16,6 +17,28 @@ namespace Bzl
             {
                 if (!callback(directory, true)) return;
                 Walk(directory, callback);
+            }
+        }
+
+        public static void PostOrderWalk(string path, Func<DirectoryInfo,bool> beforeVisitDirectory,  Action<FileSystemInfo> afterVisit)
+        {
+            var info = new DirectoryInfo(path);
+            PostOrderWalkImpl(info, beforeVisitDirectory, afterVisit);
+        }
+
+        private static void PostOrderWalkImpl(DirectoryInfo directory, Func<DirectoryInfo, bool> beforeVisitDirectory, Action<FileSystemInfo> afterVisit)
+        {
+            foreach (var sub in directory.EnumerateDirectories())
+            {
+                if (beforeVisitDirectory(sub))
+                    PostOrderWalkImpl(sub, beforeVisitDirectory, afterVisit);
+
+                afterVisit(sub);
+            }
+
+            foreach (var file in directory.EnumerateFiles())
+            {
+                afterVisit(file);
             }
         }
     }
